(#4969) Add option to control if to embed lz4 or not

+++ recipes/flann/all/conanfile.py
@@ -18,12 +18,14 @@ class LibFlannConan(ConanFile):
     options = {
         "shared": [True, False],
         "fPIC": [True, False],
-        "with_hdf5": [True, False]
+        "with_hdf5": [True, False],
+        "remove_embedded_lz4":  [True, False]
     }
     default_options = {
         "shared": False,
         "fPIC": True,
-        "with_hdf5": False
+        "with_hdf5": False,
+        "remove_embedded_lz4": True
     }
 
     _cmake = None
@@ -45,7 +47,8 @@ class LibFlannConan(ConanFile):
             del self.options.fPIC
 
     def requirements(self):
-        self.requires("lz4/1.9.2")
+        if self.options.remove_embedded_lz4:
+            self.requires("lz4/1.9.2")
         if self.options.with_hdf5:
             self.requires("hdf5/1.12.0")
 
@@ -55,8 +58,9 @@ class LibFlannConan(ConanFile):
         os.rename(extracted_dir, self._source_subfolder)
 
     def _patch_sources(self):
-        for patch in self.conan_data.get("patches", {}).get(self.version, {}):
-            tools.patch(**patch)
+        if self.options.remove_embedded_lz4:
+            for patch in self.conan_data.get("patches", {}).get(self.version, {}):
+                tools.patch(**patch)
         # Workaround issue with empty sources for a CMake target
         flann_cpp_dir = os.path.join(self._source_subfolder, "src", "cpp")
         tools.save(os.path.join(flann_cpp_dir, "empty.cpp"), "\n")
@@ -71,8 +75,9 @@ class LibFlannConan(ConanFile):
             'add_library(flann SHARED "")',
             'add_library(flann SHARED empty.cpp)'
         )
-        # remove embeded lz4
-        tools.rmdir(os.path.join(self._source_subfolder, "src", "cpp", "flann", "ext"))
+        if self.options.remove_embedded_lz4:
+            # remove embeded lz4
+            tools.rmdir(os.path.join(self._source_subfolder, "src", "cpp", "flann", "ext"))
 
     def _configure_cmake(self):
         if self._cmake is not None:
@@ -127,7 +132,8 @@ class LibFlannConan(ConanFile):
         self.cpp_info.components["flann_cpp"].libs = [flann_cpp_lib]
         if not self.options.shared and tools.stdcpp_library(self):
             self.cpp_info.components["flann_cpp"].system_libs.append(tools.stdcpp_library(self))
-        self.cpp_info.components["flann_cpp"].requires = ["lz4::lz4"]
+        if self.options.remove_embedded_lz4:
+            self.cpp_info.components["flann_cpp"].requires = ["lz4::lz4"]
         if self.options.with_hdf5:
             self.cpp_info.components["flann_cpp"].requires = ["hdf5::hdf5"]
         # flann
